+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * banners have "created" and "modified" fields. Photos have "modified" field;
 * fixed bug for cookies policy functionality;
 * added some utility links on the footer;
+* added routes for "posts of today" and "posts of yesterday";
 * improved View classes;
 * the code for backend menus and the frontend widgets has been rewritten;
 * backend topbar and backend menus now use cache.